tests: move keystore test_fixtures from bitbox02 to bitbox02_rust
`encode_xpub_at_keypath()` is going to be removed (ported to
bitbox02_rust), but `get_xpub()` will still be in `bitbox02_rust::keystore`.
